requirements.txt
# Automated Insight Engine - Dependencies
# Python 3.10+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
polars>=0.19.0
duckdb>=0.9.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Report Generation
python-pptx>=0.6.21
matplotlib>=3.8.0

# LLM Integration
google-generativeai>=0.3.0
openai>=1.3.0

# HTTP Client (for async operations)
httpx>=0.25.0

# File handling
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
